backend/requirements.txt
fastapi
uvicorn[standard]
torch
numpy
pillow
mido
python-multipart